FIX: Improve file-UI and test for issues
Improved the file-saving and opening to reduce crashes. Removed the
signals from the file_dialogs for confirmed and instead the signals for
file_selected will be used for same and better functionality. It is now
possible to double-click on a file to select it.

Found and noted some issues in the main.gd-script regarding things not
working with the file-UI. Will have to come back to these at a later
date.