FEAT: Add/remove columns and cards
This commit adds the first implementation for adding and removing both
cards and columns. It is now possible to add columns and then cards to
an empty table, as you get when starting the app.

The current implementation does need some refactor down the line,
because the future option to edit the actual table, likely needs a level
higher GUI than what is currently used. This would allow the code to
have a more natural flow also, since this is basically already used in
the card GUI.